+++ docs/app/documentation/component-docs/multi-input/examples/multi-input-example/multi-input-example.component.ts
@@ -8,5 +8,6 @@ import { Component } from '@angular/core';
 export class MultiInputExampleComponent {
 
     selected = [];
+    secondSelected = ['Pineapple'];
 
 }