+++ docs/app/documentation/utilities/api-files.ts
@@ -107,6 +107,15 @@ export const API_FILES = {
         'LocalizationEditorItemComponent',
         'LocalizationEditorMainComponent',
     ],
+    megaMenu: [
+        'MegaMenuComponent',
+        'MegaMenuGroupComponent',
+        'MegaMenuItemComponent',
+        'MegaMenuLinkDirective',
+        'MegaMenuTitleDirective',
+        'MegaMenuSubitemDirective',
+        'MegaMenuSublinkDirective'
+    ],
     menu: [
         'MenuAddonDirective',
         'MenuComponent',