+++ library/src/lib/calendar/public_api.ts
@@ -3,3 +3,4 @@ export * from './i18n/calendar-i18n';
 export * from './i18n/calendar-i18n-labels';
 export * from './models/fd-date';
 export * from './models/fd-range-date';
+export * from './calendar.component';