+++ library/src/lib/datetime-picker/public_api.ts
@@ -1,2 +1,3 @@
 export * from './datetime-picker.module';
+export * from './format/datetime-parser';
 export * from './models/fd-datetime';