+++ library/src/lib/scroll-spy/public_api.ts
@@ -1,1 +1,2 @@
 export * from './scroll-spy.module';
+export * from './scroll-spy.directive';