+++ library/src/lib/side-navigation/public_api.ts
@@ -1,1 +1,3 @@
 export * from './side-navigation.module';
+export * from './side-navigation-link/side-navigation-link.directive';
+export * from './side-navigation-sublist/side-navigation-sublist.directive';